requirements.txt
kiwixstorage>=0.8.1,<1.0
pif>=0.8.2,<0.9
zimscraperlib>=1.6.1,<1.7
xml_to_dict>=0.1.6,<0.2
cli-formatter>=1.2.0,<1.3
py7zr>=0.16.1,<0.17
python-slugify>=5.0.2,<6.0
jinja2>=3.0.1,<3.1
redis>=3.5.3,<3.6
beautifulsoup4>=4.9.3,<5.0
lxml>=4.9.1,<4.10
jinja2-pluralize>=0.3.0,<0.4
tld>=0.12.6,<0.13
mistune>=2.0.0rc1,<2.1
python-dateutil>=2.8.2,<2.9
psutil>=5.8.0,<5.9
python-snappy>=0.6.0,<1.0
bidict>=0.21.2,<0.22
cchardet>=2.1.7,<2.2
